ScarceNet-main-original/ScarceNet-main-original/requirements_wout_torch.txt
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
Cython==0.29.24
dataclasses==0.6
easydict==1.7
future==0.18.2
idna==3.3
imageio==2.10.1
imgaug==0.4.0
joblib==1.1.0
json-tricks==3.15.5
jsonpatch==1.32
jsonpointer==2.2
kiwisolver==1.3.2
matplotlib==3.4.3
networkx==2.6.3
numpy==1.20.2
opencv-contrib-python
opencv-python
pandas==1.3.4
Pillow
progress==1.6
protobuf==3.19.1
pycocotools
pyparsing==3.0.4
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
scikit-image==0.18.3
scikit-learn==1.0.2
scipy
seaborn==0.11.2
Shapely==1.8.0
six==1.16.0
tensorboardX==1.6
threadpoolctl==3.1.0
tifffile==2021.10.12
torchfile==0.1.0
torchnet==0.0.4
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.0
urllib3==1.26.7
visdom==0.1.8.9
websocket-client==1.2.3
xtcocotools==1.10
yacs==0.1.8
